Create new module 'execution' and move all top-level files into a module (commands, parsers, execution)
This just leaves tabcmd.py as the entry point
