+++ tabcmd/execution/map_of_commands.py
