+++ tabcmd/execution/map_of_parsers.py
@@ -1,30 +1,30 @@
-from .parsers.create_project_parser import *
-from .parsers.login_parser import *
-from .parsers.get_url_parser import *
-from .parsers.delete_project_parser import *
-from .parsers.create_group_parser import *
-from .parsers.delete_group_parser import *
-from .parsers.create_users_parser import *
-from .parsers.remove_users_parser import *
-from .parsers.export_parser import *
-from .parsers.logout_parser import *
-from .parsers.add_users_parser import *
-from .parsers.create_site_users_parser import *
-from .parsers.create_site_parser import *
-from .parsers.delete_site_parser import *
-from .parsers.delete_site_users_parser import *
-from .parsers.edit_site_parser import *
-from .parsers.list_sites_parser import *
-from .parsers.delete_parser import *
-from .parsers.publish_parser import *
-from .parsers.create_extracts_parser import *
-from .parsers.decrypt_extracts_parser import *
-from .parsers.delete_extracts_parser import *
-from .parsers.encrypt_extracts_parser import *
-from .parsers.reencrypt_parser import *
-from .parsers.refresh_extracts_parser import *
-from .parsers.publish_samples_parser import *
-from .parsers.help_parser import *
+from tabcmd.parsers.add_users_parser import *
+from tabcmd.parsers.create_extracts_parser import *
+from tabcmd.parsers.create_group_parser import *
+from tabcmd.parsers.create_project_parser import *
+from tabcmd.parsers.create_site_parser import *
+from tabcmd.parsers.create_site_users_parser import *
+from tabcmd.parsers.create_users_parser import *
+from tabcmd.parsers.decrypt_extracts_parser import *
+from tabcmd.parsers.delete_extracts_parser import *
+from tabcmd.parsers.delete_group_parser import *
+from tabcmd.parsers.delete_parser import *
+from tabcmd.parsers.delete_project_parser import *
+from tabcmd.parsers.delete_site_parser import *
+from tabcmd.parsers.delete_site_users_parser import *
+from tabcmd.parsers.edit_site_parser import *
+from tabcmd.parsers.encrypt_extracts_parser import *
+from tabcmd.parsers.export_parser import *
+from tabcmd.parsers.get_url_parser import *
+from tabcmd.parsers.help_parser import *
+from tabcmd.parsers.list_sites_parser import *
+from tabcmd.parsers.login_parser import *
+from tabcmd.parsers.logout_parser import *
+from tabcmd.parsers.publish_parser import *
+from tabcmd.parsers.publish_samples_parser import *
+from tabcmd.parsers.reencrypt_parser import *
+from tabcmd.parsers.refresh_extracts_parser import *
+from tabcmd.parsers.remove_users_parser import *
 
 
 class ParsersMap:
