+++ tabcmd/execution/tabcmd_controller.py
@@ -1,8 +1,3 @@
-import argparse
-import sys
-from .parsers.parent_parser import ParentParser
-from .logger_config import get_logger
-from .parsers.help_parser import HelpParser
 from .map_of_commands import *
 from .map_of_parsers import *
